Save conversation history, update API usage
This adds support for setting the isArchived bit, which causes groups to
retain their conversation history in the Groups UI.  This feature is
part of a separate service, so it requires a new scope, which will
invalidate the previous OAuth token and require you to log in again.

This change also updates the API to replace the deprecated call to New()
with the shiny new NewService().  This supports simpler auth options
(like default credentials), but I override that by passing in the
http.Client from the googoauth library.