Make the script much more forgiving
This adds retries if the group doesn't immediately exist when we try to
add the primary member, and does not "fail" if the list already exists,
or has the described members in it.  This enables you to use the script
to "set" the current state of the code on any "existing" list, or to
just generally retry-on-error.